+++ src/client.ts
@@ -1,6 +1,11 @@
-export { fromErrorResponse, parseErrorResponse } from './error-response-data';
+export {
+  fromErrorResponseData,
+  parseErrorResponseData,
+} from './error-response-data';
+export { fromHttpResponse } from './from-http-response';
 
 export type {
   ErrorResponseData,
-  FromErrorResponseOptions,
+  FromErrorResponseDataOptions,
 } from './error-response-data';
+export type { FromHttpResponseOptions } from './from-http-response';